requirements.txt
fastapi
uvicorn
google-cloud-bigquery
pandas
python-dotenv
requests
db-dtypes
pytest
httpx
